Refactor : Canvas.jsx 기능별 역할 분리 및 훅 정리
